Add option driver
The option driver reads custom proto options defined in the gen/svc/annotations
file. Choices in this driver were deliberately made to be a drop in replacement
to the comment driver. Numbers are convereted to strings and functions return nil
because there is no error to return. This should be revisited to once
the comment driver is completely removed.